"10am. Last night I played for too long. I went to bed past 12am (again) but I really do not get as much time to myself as I wanted. Even though I stopped at 5pm once bath and lunch was over somehow it was 7pm. Once I finished reading manga it was 8pm. There is really not much room to fit play in there unless I go to bed later than the usual 9-10pm.
Demonsteele + Scythe 2 is really great. It gets me into the demon-killing zone like nothing else. Why aren't real life jobs like that? There is something wrong with this world or the brain, something that I want to fix.

I am in danger. Since nobody can wrench the field free from backprop, that means I have to take it up myself and I am not sure how long will I have leeway to learn and do research.

Sure I want to be a programmer, but not the kind of programmer that serves. I want to be a warlike figure - I want to use my powers of reasoning to defeat other people. That is what I have decided.

https://www.reddit.com/r/MachineLearning/comments/bb9umg/d_my_machine_learning_research_job_interview/

Get a PhD in a hot field and then get the companies to pay out their ass for your services - I am not sure this is who I want to be. I'd rather make 5k in my own way than 200k like this.

The companies are definitely overpaying for data scientists at this juncture. It is not like they will be causing the Singularity. I mean it is true that data science requires specialized skills, but the reason why the whole field is hot - namely deep learning requires the least expertise. Deep learning is all about potential. But it is not like potential itself has that much value.

I can only shake my head at this.

If I find this frustrating, I should spend that energy into improving myself.

10:25am. Mhhh, I really did sleep well tonight. Let me chill just a bit more and then I will start this.

I've mastered the fundamentals of programming dependent types - the fundamentals of machine learning have been long overdue."