"10:45am. Let me finally start. I'll do this for 30m and then have breakfast.
11:40am. 70/748. Let me stop here. I can't finish the chapter right now. Since I had not had breakfast last night, I am really starving here.

12pm. https://arxiv.org/abs/1904.04025
Samples are not all useful - Denoising policy gradient updates using variance

I'll leave this for later.

12:30pm. I've resumed.

71/748. Somehow, I can follow the stuff here. Not anticipate, but follow.

Hopefully I will get something out of it by the time the book is done.

12:45pm. Focus me, focus. Stop surfing 4chan on the side. Let me just get through the chapter.

1pm. 76/748. Ok, this is it for chapter one.

Now these exercises - for once I should actually do all of these...except I really am not good enough at math to really grok this just yet.

1:10pm. I need to make a determination to go through all of this, but rather than do it as an exercise, I should take this opportunity to put my programming skills to bear.

1:20pm. Well, of course that will come later. First I need to go through the book once.

After that I will find a way to have Idris assist me when I am doing those problems. Maybe there are different kinds of math tools that could do this as well.

1:25pm. What I am trying to do here is make a change of mindset. Rather than looking at these exercises as a burden, I need to start looking them as an essential component of the journey. These exercises are the real meat of the book.

What I need to do here is go beyond pen-and-paper methods that have been used since antiquity. This is the true goal of going through the ML book. I mean even if I do go through it - there is no reason to think that I will suddenly become capable of dealing with the credit assignment problem. But if I can abstract the book itself then that would be a massive achievement that would separate the me of the present from the me of the past.

It would be the Spiral language all over again.

All the challenges ahead of me are in essence language problems. The ways of thinking and the ways of expression in the present are not enough and I need novel kinds of abstractions to deal with the challenges.

1:35pm. How could I not just do the exercises, but make them an asset to the programs I am making?

That is the true purpose of having advanced type systems. They allow one to draw out the essence of programming.

Let me take a break here as I need to do the chores. I'll start chapter two when I get back."