"2:20pm. Since I've made the determination of what I want to do, I am thinking of skimming the next 100 pages and then starting on the project.
https://en.wikipedia.org/wiki/Computer_algebra_system

Actually, I never seen a real CA system. I had no idea there are courses on them. Maybe having a look into them might be worth a try.

https://www.wolfram.com/mathematica/trial/

Ugh, 15 day trial. This stuff is cancer.

https://www.ted.com/talks/conrad_wolfram_teaching_kids_real_math_with_computers?language=en

Watching this talk might be worth the time.

http://www.wolfram.com/mathematica/

To be honest, Mathematica's capabilities look amazing.

http://maxima.sourceforge.net/index.html

But I want something open source to start with. I want to figure out what these things are all about.

2:35pm. Let me watch the TED Talk by Conrad Wolfram. Is that guy Steven's brother?

3:05pm. Oh damn it. Let me get back to the book.

The talk really spoke to the heart. Computers are definitely the next step in the evolution of mathematics.

3:25pm. 97/748. "However, we strongly encourage the reader to become proficient in manipulating Gaussian distributions using the techniques presented here as this will prove invaluable in understanding the more complex models presented in later chapters."

Yeah, I am at my limit here. It is like the book is saying to me that I should stop just reading it and start doing math.

I really need to master a CAS system here. This is the most sensible next step.

I had it all wrong. I do not know when it was that I got this idea, it must have been in elementary school what I was doing Basic and Logo, that computers are good for simple calculations and useless for symbolic manipulation. I have plenty of evidence in the last 4 years to the contrary for that. I need to tap into that power. The odds that I am going to be able to do this are low enough as they are - the last thing I should do is vastly increase the possibility for error by doing calculations by hand.

I will do it all as a programmer.

From here on out, I'll try the book from the start, this time using Maxima to type in the various formulas and see where they lead me to.

Before that, let me install the programming language and go through the documentation for it.

That is the most essential first step."