"8:35am. I went to bed earlier than usual yesterday night, but that was probably a mistake as I kept being wracked by worry until I feel asleep. Even today, I must have first woken up hours and ago and spent this time in worry again.
I have a lot of to be worried about. Trying to tackle math itself as a language project in order to find the algorithm I need seems like a fairly insane undertaking. Beyond mastering fundamentals lies the computational space that I need to mine. Deep learning is the wings of AI, but it is also a prison that keeps confines to areas below the atmosphere. As a framework it allows a lot of freedom with regards to models, but none of the directions it allows evolution in can result in resolution of the credit assignment problem. Deep learning does not allow true hierarchy. That is the issue.

It is unfair. The easy way was always to go with what exists. The sensible thing was too look for a job in 2015 and have my employers pay for my education in whatever I was doing.

The hard thing is to unveil the next step. It is not just hard - solving the credit assignment problem is pretty much impossible for me. Even if I could at the very least it will take me a few years, but it is the kind of thing that could stretch for well over a decade. I can't do this forever, but while I have the vitality in me, I need to try it. Clearly somebody crazy enough to actually attempt this is needed.

I cannot tackle it directly, but I can stitch together pieces as they come along.

There is a kind of convergence going on with advanced type systems and program synthesis. The same goes for computer algebra and probabilitic programming. I've never seen symbolic regression scaled to a grand level either.

8:50am. Deep learning feels like a buggy program.

How do I put it? When one writes a program, one does his best to make sure that it works right the first time around. He places all his hope and trust in that. Those hopes are then dashed invariably as there are runtime and type errors and there is a period of fixing the initial blueprint.

I sort of got the sense that this is how my attempt at making deep nets work for RL would go in 2018.

Yeah, of course I would fail. But the difference between trying and failing and not trying at all is like the difference between having an empty file and an almost complete program.

I could see that I would fail, but I could not see the conclusions I would reach after that happened. Of the grand papers that I've gone into : Zap, Rudder, KFAC and differentiable plasticity, rather than my reasons for trying them out being dead, my programmer's intuition is telling me that backprop itself is in the wrong.

With Rudder, moving from gradients to contribution percentages is a significant conceptual leap. And differentiable plasticity is a blueprint for supervised systems being built on top of unsupervised ones. That is not going to go away. Zap, even though it is shallow does not depend on gradients has fairly intricate reasoning that impressed me and KFAC might be an omen that backprop could evolve through better optimizers. Before KFAC, I thought second order optimizers were hopeless.

9am. Squeezed in by the impossibilities where am I going to find an exit? How will I deal with the contradiction of wanting money and yet being pushed into research instead? There is no point in answering those questions.

All I can do now is master the fundamentals of ML.

In hindsight, I think the way I've been taught math in school has harmed me greatly. If math should be 100% done on the computer, then the ingrained habits have been setting me up for failure.

Math is the beginning and the end of everything. I've always been bad at it and unless I surmount those flaws with the power of machines, I will always be lagging behind the researchers who have inborn talents for it.

If I can do this properly, a great power will be born.

9:15am. Writing the above has shaken off my grogginess a little. Let me do my morning reading and then I will start things off by installing Maple. I want to see what it has. Also I've never finish the talk on program synthesis that was posted on the PL sub. I should get to it as well.

Let me do it. Even though this might take a decade, I am not allowed a day off work. This is the life I've chosen for myself."