"11:45am. Done with chores. Let me finally start. Time to install I looked into Maxima's manual yesterday and it is too much of a mess for me to really accept it. Maple and Mathematica should be more promising.
11:55am. Finally the install is done. Let me finally try running it.

Ok, it runs. Somehow Maple brings to mind the old VS experience back in high school. I think I pirated VS over 15 years ago and never did anything, but I remember having MSDN on a stack of CDs. Like then, there is a wealth of documentation right there on my fingerips. It is obvious even at first glance that Maple is worlds apart from Maxima.

12:15pm. There is a lot of stuff here, quite neatly arranged. I am quite impressed. In the worst case, I will never have to write math using pen and paper again. Or using standard text editors.

...It is cool how it is possible to mix text and math right there in the worksheet.

And visualizations and standard controls.

1pm. Ok...

I was engrossed in this for a bit, but now it is time that I try and do the exercises from the ML book.

This is how math should be done. I really should not be using pen and paper in 2019, nor doing it in a text editor. And if I had to do things like simplify expressions, it would be a better idea to just write a compiler for that.

1:10pm. 76/748. The exercises start here.

Now what is the polynomial given by 1.1?

1.2 is the squared error divided by two.

1.1 is : `inl y (x,w) = sum (from: 0 to: M) (inl j -> w j * x ^ j)`

When writing math in text in the journal here, I am going to use Spiral's syntax for notation.

1:15pm. Now let me implement this function in Maple somehow. I'll figure out what the problem is asking me after that.

There should be something in the documentation regarding polynomials. I am unsure of how to index into a variable. Also what will I define M to be? Obviously the size of `w`..."