I had not intended to play Brutal Doom till 1:30am yesterday, but I did somehow. I've been trying out various mods and watching the mod madness vids, but they are all gimmicks. I did not expect much from BD given that I've played it to death years ago, but damn is the latest version good. It is amazing. The gameplay is worlds apart from say Insanity's Requiem whose flashiness tricked me. I should not have wasted my time on the other mods. Demonsteele is the notable exception.
Now, with that little rant out of the way, let me start the day. I've thought about it for a bit and this is the ideal spot to try out Mathematica now that I've gotten a tad familiar with Maple.

Maple is extremely competently done, but its language design leaves much to be desired for me. Maybe Mathematica will be more to my taste?

That is what I will be finding out today.

I already have it sitting on my drive, so let me try it out.