"11:45am. http://www.wolfram.com/language/fast-introduction-for-programmers/en/function-definitions/
This is impressive stuff. I am reminded of the old saw that a language that does not change your view of programming is not worth learning. Mathematica definitely has this going in its favor over Maple.

"You can define transformations for any left-hand side—a convenient way to "destructure" arguments:"

Pattern matching. That is what I like. I think that I will go with Mathematica if nothing comes up to make me change my mind which I expect won't.

12:35pm. http://www.wolfram.com/language/fast-introduction-for-programmers/en/strings/

I managed to get this far. This kind of language design is definitely much more to my taste.

12:40pm. Focus me. Finish the tutorial. Stop browsing 4chan.

1:10pm. Done with the quick tutorial.

http://www.wolfram.com/language/elementary-introduction/2nd-ed/preface.html

Stephen Wolfram's book might be worth going through. I guess I will do it. I quite like the overall design of Mathematica. Pattern matching everywhere is my kind of thing.

1:20pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/what-is-the-wolfram-language.html

I guess I'll dedicate the next few days to this book. This will give me some basic exercise in the language itself.

1:50pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/04-displaying-lists.html

Plotting and visualization is just so easy in Mathematica. Doing this would just be so much pain in the ass in any other language I've looked at. I am specifically looking at F# here.

http://www.wolfram.com/mathematica/pricing/home-hobby-individuals.php

Like Maple, Mathematica is damn expensive at 225 pounds per license. I am not going to be getting a license for this anytime soon.

2pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/06-making-tables.html

This book is super easy. I will be done with it today most likely. Unless I get distracted and start browsing the web randomly...

2:10pm.

```
Table[RandomInteger[10], 20]
```

This computes random integers separately. Interesting.

`Table` is basically `init` anywhere else.

2:20pm. Let me take a break here. I need to handle the chores.

4pm. This sure took a while. Let me put in an uninterrupted hour and a half into this. I should get at least past halfway into the book.

4:20pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/10-images.html

This stuff is quite impressive. When I said wanted GUIs, this is what I really wanted the most out of it.

5:10pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/19-dates-and-times.html

"If you know both the date and a location on Earth, you can work out when the sun will rise and set."

```
Sunset[Here, Today]
```
```
DateObject[{2019, 4, 14, 19, 40}, "Minute", "Gregorian", 2.]
```

The output is ugly here, it is rendered smoothly in the IDE.

5:25pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/21-graphs-and-networks.html

I really should go on for another hour, but I am deeply bored by this. Let me give it till chapter 25 and then I will call it a day here. I want to have some fun. Hopefully after a few more days, I will both be done with this boring tutorial stuff and have gotten tired of games. The will allow me to properly focus on ML. The day I get those chapter one exercises out of the way will be the day I will have made another step forward as a person.

5:45pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/26-pure-anonymous-functions.html

Let me stop here. I am going to flip out if I have to bear anymore of this boredom. My focus is low today. This stuff is super easy, but I have to go through it methodically and it is making me fall asleep at the wheel here.

Tomorrow I will have a basic grasp of Mathematica's features and will be able to move on to the stuff I am actually interested in. There should be some stuff in it on least squares. I'll still be doing tutorials for the foreseeable future, but I should be able to start making progress on the first exercise from Bishop's ML book.

It is not enough that I just solve that exercise - I need to delegate the solution to the computer as much as possible. I need to do that for the entire book. The book is just standard statistics stuff. I need to abstract the material, move it to the computer and so master it.

I do not care if it takes me an entire month. If I could do this, then I will have closed a significant part of the gap between me and the PhD researchers who I am trying to defeat.

I can only do this one step at a time.

I'll do this for as long as the circumstances allow."