I've been having way too much fun lately. 2019 so far has been a one big case of massive downtime for me.
It is not even that I am being lazy - it is more like that going beyond simply implementing various algorithms is way too hard for me at the moment, so all I've been doing is learning in an unfocused manner. I've been wandering about and wasting too much time as a result.

The gap between what I want to do and what I am capable of doing just won't go away, not so easily. There is no bargaining with the universe. The only choice was always to do it right.

I need to move beyond being a programmer into being the programmer. For that I need to draw in all the skills that there exist.

As I mentioned, I've been having fun playing Doom and reading manga. I remember a time when those things weren't a timewaster.

There was a time where not only I played, but I also believed. The games weren't just another thing - they are the defining trait and the pinnacle of culture. I could not understand why normies could not see it.

But the last 15 years were not kind to their development - the graphics might have gotten better, but they are being help back by human ingenuity.

The worst of it would be - they are a time waster. I imagined a time where one day the games would be a source of power to those who play them. The feeling I got was certainly like that.

I will never forget nor forgive those looks as if I was some kind of bug stuck in a trap made to weed out useless people. I will make my perspective the correct one.

I won't let it end at this. The games weren't created to weed out certain classes of people from the gene pool.

They were created to nurture talent.

When games become power, the Singularity will manifest itself and the era of man will come to an end.

I must never forget what I am trying to do.

I've done the impossible before and I need to do it one more time.

I can't be held back by something as stupid as math. By any reasonable estimate, the programming I've been doing is way more complex than math itself.

That I am so bad at it is some universal absurdity that I need to get to the bottom of. There is no doubt that the solution will be to apply computers.

Once I figure out the fundamentals of both ML and algorithmic analysis, I will be able to make progress on the whole credit assignment puzzle and make headway on debugging the issues surrounding backpropagation.

This can be done. Eventually there will be limits on what can be expressed when it comes to machine learning, but I cannot imagine that what currently exists is that limit.

What I am able to say now cannot possibly be the limit of my ability.