"12pm. ">>187104471 It is pretty interesting. The kinds of developments in this manga are not something an ordinary writer could do.
Some writers like Fukumoto or One have a firm grasp on humanity. Murata on the other has a solid grasp on the more inhuman, subconscious drives that lurk below the ego.

It only seems dumb if you use the standard reality as a reference, but if you think about some of the characters being a different species then it clicks into place and you can have some fun with this story.

Rather than seeing this as a tragic backstory, I see this more as highlighting the girl's inhumanity. Rather than being crushed by the despair which an ordinary human would have been, she made it her snack. She liked the kind of life that she had. If you think about it what she is doing now - she is essentially going on a rampage not because of abuse, but because the abuse stopped. Peace is poison to her and she thrives in an environment of struggle, both inner and outer.

It is not tragedy, but a way of signaling that Mizu is one of the next steps in the evolution of humanity."

I am the one writing this particular reply. By now, I changed my mind on anybody other than Himeno being able to convert other people, but I think this is what the author is driving at. Now I hope my IP range is unblocked. Some asshole fucked it up last time. Ah, great it is not.

12:15pm. Let me resume the book. I am done with breakfast. Hopefully I will finish it today. As usual despite any kind of determination, the going is slow.

It is fine. I can get through this super basic tutorial book even 10 year olds could grasp.

```
NestList[Join[{0}, #] + Join[#, {0}] &, {1}, 8] // Grid
```

This is a really interesting way of doing the Pascal's triangle. I had not known it was possible to do it as such.

12:40pm.

```
FoldList[f, x, {1, 2, 3, 4, 5}]
```

These symbolic languages are just wonderful.

```
{x, f[x, 1], f[f[x, 1], 2], f[f[f[x, 1], 2], 3],
 f[f[f[f[x, 1], 2], 3], 4], f[f[f[f[f[x, 1], 2], 3], 4], 5]}
```

Has fold ever been explained in such a direct manner before? This might be the best explanation ever.

http://www.wolfram.com/language/elementary-introduction/2nd-ed/29-more-about-pure-functions.html

I'll admit, by this point I am wondering how the language will handle nested anonymous functions. They are a thing.

1:10pm. https://www.maplesoft.com/support/help/Maple/view.aspx?path=examples%2Fpatmatch

Maple does have pattern matching, but it is a bit different than I imagined it would be.

Also before I jump to conclusions about it, I need to get familiar with its basic programming model myself. I haven't been thorough with respect to Maple. I still have a while to go before I could form a conclusive opinion on either language.

1:45pm. https://blog.stephenwolfram.com/2018/06/weve-come-a-long-way-in-30-years-but-you-havent-seen-anything-yet/#more-17068
"Thirty years is basically half the complete history of modern digital computing."

30 years is nothing. It cannot be the end, not so easily.

2pm. Focus me. Get back into the book.

2:20pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/36-creating-websites-and-apps.html

What tedium, but I am 3/4ths of the way in. Let me finish it and then I will take a break.

2:25pm. Let me skim this chapter. I am doing that anyway, but I do not dare to log into the cloud. Mathematica keeps pestering me for my ID, but as a matter of principle I do not want to do it.

2:35pm. "x++ is equivalent to x=x+1. x+=n is equivalent to x=x+n. AppendTo[x, n] is equivalent to x=Append[x, n] or x=Join[x, {n}]."

So Mathematica has the C increment and decrement. Ok, in that case I won't hold it against Maple for adding them.

2:45pm. It is barely 3pm and I am already getting lost in thought.

http://www.wolfram.com/language/elementary-introduction/2nd-ed/39-immediate-and-delayed-values.html

This is something I should pay attention though. It is interesting.

3:10pm. Done with chores. Focus me. I need to resume this.

Also I've been thinking of giving Coq a try again, or the very least I should go through the SF exercises in Idris.

3:25pm. `plusminus[v_] := {v[[1]] + v[[2]], v[[1]] - v[[2]]}`

Ah, I was wondering about the weird list syntax, but it turns out the syntax is specifically for variable arguments.

3:40pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/41-more-about-patterns.html

This is easy, but I am having extreme difficulty focusing on this. Nonetheless, I get it.

3:50pm. Mathematica's pattern matching is more powerful than in the ML family of languages. I quite like it.

I am going to have to keep this chapter in mind.

4pm. I need a break. This is killing me. I'll put in some extra effort and do the last 7 chapters after that. Damn this.

4:25pm. ">>210018195
The body and the soul were never really together and I doubt they even existed as a concept before a certain point - religion's great innovation was to invent the soul. Science just did not get around to dealing with it yet, mostly because it did not need to. But it killed most of the old elemental gods and put them into metal, now it needs to get around to killing the newer ones and putting them into code. Kill the man behind the God and whatever remains will be power, a ready and usable abstraction.

The disease called the soul will be cured once technology gets to the point where it becomes possible to control matter at such a level that intelligence augmentation becomes possible.

Knowing what intelligence actually is and being able to modify minds directly will also get rid of the whole IQ nonsense as well. Rather than have rough paper tests, it stands to reason that it should be possible look at a mind directly and estimate intelligence.

Of course, actually being to do the above will mark the end of the era of man. The hardships of the self improvement process will fuse the body and the soul, but it will also pretty much kill everybody and everything in the process.

It won't be end of life, but a fundamental change to the very conception of what life actually is.

To cope with that, a new kind of religion will be needed. This has historically been the solution to matters of life and death - to believe in a falsehood from which everything flows.

Since everything ends in falsehood, for the next iteration of religion, it might be better to personalize it rather than be forced to maintain a collective delusion to minimize the friction on society and cut back on the bullshit.

...Too bad about normies though. No way will they survive the death of both God and soul in any kind of form. But well, dinos had to die to make room for humans, it is their turn next."

Now this is just a shitpost. The above should be true, but we already have a sort of embrionic brain in the form of NNs and I can't do damn with them. It would be just too damn embarrassing to say this out loud in the light of that. There is no point in prancing around like a clown and trying to enlighten other people - it would be best to complete the project and finally weaponize the insights. Then things will fall in line. That is the whole point of the programming journey.

I am not going to post this as a reply.

I must be crazy today. I really will be lost, if I start posting on /pol/. I already made one dumb post on the Himenospia thread, no way am I going to make that a pattern. Himenospia really deserves it - Murata Shinya is a lot better writer than people give him credit for.

4:30pm. Ok, the last spurt. I will finish the book from here on out and call it a day.

Well, before calling it a day I will do a short survey of the documentation just to see where curve fitting stands.

4:35pm. Focus me. Do this thing and you will be able to get back to slaying demons."