"5pm. Hmm, it is this late already? Where did the last 30m go?
I spent some time of that multiplication commutativity proof that I did in Idris. A certain thing confused me, but it I now understand why it simplifies that way it did.

Yeah, I get the thing.

Let me finish the book.

5:10pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/46-writing-good-code.html

I still haven't gotten an answer on how nested anonymous functions work.

https://mathematica.stackexchange.com/questions/16947/second-level-depth-pure-function

Ah, it has proper formal `Function`s as well. Ok, that takes care of that.

5:25pm. http://www.wolfram.com/language/elementary-introduction/2nd-ed/what-we-havent-discussed.html

Ah, finally. I am nearly done. Just 3 more chapters left.

5:35pm. Oh, there is stuff on mathematical computation here. This is exactly what I need. I will take note of this.

...And I am done. I have very basic grasp of the language now. It took me 2 days, but I did it.

5:40pm. I'll leave the survey for tomorrow. I do not feel like even more documentation for one day. I have a low tolerance for this stuff.

Tomorrow I will be able to start ramping up. I need to get that ML book problem done and I'll give doing it in Mathematica a shot. After that I'll see how Maple does it. By this point I've pretty much forgotten what I was doing in Maple so I'll have to refresh my memory.

Ah, yes trying out the regression analysis stuff.

I remember being confused by all the options regarding plotting that Maple has, but it is not any worse than Mathematica. The two languages have fairly similar designs overall.

5:45pm. There is no any way around it. I just have to roll up my sleeves and get busy from this point on.

With regards to Coq, I decided to drop it because of the strange induction behavior, but today I've come to the conclusion that even though it is strange, I do not really have to worry about anything else.

Really, induction is the one thing that I need to understand to grasp DT programming and now I have it. It is all about various kinds of rewrites after that. I can anticipate how things should be now.

I'll need the experience later on.

Let me focus on this. 2019 will be a year of self improvement.

The ultimate goal is to come to an understanding of how ML algorithms should be constructed from the ground up.

If I manage to attain that - this world is finished."