I am sort of getting back into the more normal frame of mind after all the debauchery of the last week.
Hmmmm...

It is going to be a pain in the ass, the things that are coming up next. But I have to do it.

The most important thing if I am going to be a researcher will be to master the fundamentals. There are thousands of papers coming out every year and literally none of them are worth reading.

Those benchmark papers are useless.

They won't get me any closer to my goal.

The more I think about the more pissed I get at the Deepmind/OpenAI media circuses. Yeah, it is is progress to know what PPO is capable of, but being able to beat Dota and Starcraft won't give any insight as to what is needed to make a much better algorithm.

Power. How I regret not having it. Just to what lengths does one have to go to reach it?

I've been doing this for 4 years and it is still not enough. Just what does one have to do to even get computers to the toddler stage?

I regret it. I though I would be able to be an implementor, but just that is not feasible. There is so much disconnection with reality where that desire is concerned.

If the general understanding of ML were good, then expertise in implementation especially to the lengths I went to with Spiral would be a first rate ability, but right now I am barely a second stringer. It is useless.

That having said, it makes sense.

Making Spiral makes sense in the grand scheme of things. It makes sense to start with the low level and then move on after that.

I'd have preferred not to have to spend so much time on it and even though it is not at all what I need, it makes sense. Not being able to implement standard deep learning algorithms on my own would be an embarrassment as a practitioner. If something something else came along that means I would need to wait for Google and Facebook to get around to making a library for it.

Yet, I regret spending more than a few months on Spiral.

And right now I regret that I won't be able to go into poker. I regret that it will take me years to get good at algorithm design.

This might very well fail. I might spend years and then decades of my life going nowhere.

If that happens I will regret it at the end of it.

It is time to start. In order to avoid the death that I will come in the decades hence, I will face it in the present instead. Let that perspective become the fuel for what I will do from here on out.

Let me investigate Mathematica's capabilities for mathematical computation.

http://www.wolfram.com/language/elementary-introduction/2nd-ed/what-we-havent-discussed.html

It is the stuff that is here that is of interest to me. Let me go to it. Once I solve the first problem I will, get familiar with how the same thing is done in Maple.