"10:10am. Damn that took a while. Let me continue going through the documentation. These commits are long and prosaic, but they build motivation.
10:15am. http://www.wolfram.com/language/fast-introduction-for-math-students/en/

Let me go through the intro for math students. This is what I need here.

11am. http://www.wolfram.com/language/fast-introduction-for-math-students/en/limits/

Let me stop here no, so I can have breakfast, but so I can read the Dijkstra's quotes.

https://www.cs.utexas.edu/users/EWD/transcriptions/EWD08xx/EWD898.html
https://www.cs.utexas.edu/users/EWD/transcriptions/EWD10xx/EWD1036.html

11:15am. "On the historical evidence I shall be short. Carl Friedrich Gauss, the Prince of Mathematicians but also somewhat of a coward, was certainly aware of the fate of Galileo —and could probably have predicted the calumniation of Einstein— when he decided to suppress his discovery of non-Euclidean geometry, thus leaving it to Bolyai and Lobatchewsky to receive the flak. It is probably more illuminating to go a little bit further back, to the Middle Ages. One of its characteristics was that "reasoning by analogy" was rampant; another characteristic was almost total intellectual stagnation, and we now see why the two go together. A reason for mentioning this is to point out that, by developing a keen ear for unwarranted analogies, one can detect a lot of medieval thinking today."

"For instance, the vast majority of the mathematical community has never challenged its tacit assumption that doing mathematics will remain very much the same type of mental activity it has always been: new topics will come, flourish, and go as they have done in the past, but, the human brain being what it is, our ways of teaching, learning, and understanding mathematics, of problem solving, and of mathematical discovery will remain pretty much the same. Herbert Robbins clearly states why he rules out a quantum leap in mathematical ability:"

""Nobody is going to run 100 meters in five seconds, no matter how much is invested in training and machines. The same can be said about using the brain. The human mind is no different now from what it was five thousand years ago. And when it comes to mathematics, you must realize that this is the human mind at an extreme limit of its capacity.""

"My comment in the margin was "so reduce the use of the brain and calculate!". Using Robbins's own analogy, one could remark that, for going from A to B fast, there could now exist alternatives to running that are orders of magnitude more effective. Robbins flatly refuses to honour any alternative to time-honoured brain usage with the name of "doing mathematics", thus exorcizing the danger of radical novelty by the simple device of adjusting his definitions to his needs: simply by definition, mathematics will continue to be what it used to be. So much for the mathematicians."

11:25am. "Because, in a sense, the whole is "bigger" than its parts, the depth of a hierarchical decomposition is some sort of logarithm of the ratio of the "sizes" of the whole and the ultimate smallest parts. From a bit to a few hundred megabytes, from a microsecond to a half an hour of computing confronts us with completely baffling ratio of 109! The programmer is in the unique position that his is the only discipline and profession in which such a gigantic ratio, which totally baffles our imagination, has to be bridged by a single technology. He has to be able to think in terms of conceptual hierarchies that are much deeper than a single mind ever needed to face before. Compared to that number of semantic levels, the average mathematical theory is almost flat. By evoking the need for deep conceptual hierarchies, the automatic computer confronts us with a radically new intellectual challenge that has no precedent in our history."

This whole essay so wonderful. Rather than just a few paragraphs, the two essays by Dijkstra are worth reading in full.

There is zero reason for me to be bad at math considering how good I am at abstract thinking. Decades?

Yeah, right. Once I figure out how to point my talent at math, it will fall just like anything else.

12:35pm. As soon as I finish catching up with Asuka I will resume. I did not think that the essay would take me nearly an hour.

1:25pm. Done with chores. Seriously, I need to get going.

2:15pm. Focus me, go through the math student guide. Stop surfing /pol/ on the side.

2:30pm. http://www.wolfram.com/language/fast-introduction-for-math-students/en/probability/

I can expect Maple to have all this functionality as well.

2:45pm. http://www.wolfram.com/language/fast-introduction-for-math-students/en/data-plots-and-best-fit-curves/

Here is the curve fitting stuff.

http://www.wolfram.com/language/fast-introduction-for-math-students/en/math-puzzles/

Though not directly, a lot of the stuff I am going to be doing would fall into the puzzle category.

3:05pm. Let me stop for a bit now that I've gone through the Fast Introduction For Math Students. I need to curate my thoughts.

3:10pm. With this I feel like I've seen enough of Mathematica to have a decent overview of its fundamentals.

Of course this is not really enough to be anywhere near proficient at either Maple or Mathematica, but the languages are no longer impenetrable. I'll fill in the holes in my knowledge as I go along.

At this point for the next month or two, all I should do is concentrate on going through the exercises for the ML book.

Now...what was that first problem on page 76 again? I'll visualize it in all its forms for small `n`s, solve for those and then do it for an arbitrary number of weights.

I had a little taste of math while I was doing those rewritings for KFAC, but this will be the first time in my life that I will take up the mantle of mathematician out of my own will. I won't fail with the power of machines at my side."